refactor(date-entity): DateEntity 추상 클래스 활용해 각 entity의 createdAt, updatedAt 컬럼 추상화
