+++ src/main/java/com/shootdoori/match/entity/MatchApplication.java
@@ -1,16 +1,21 @@
 package com.shootdoori.match.entity;
 
-import jakarta.persistence.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "match_application")
-@EntityListeners(AuditingEntityListener.class)
-public class MatchApplication {
+public class MatchApplication extends DateEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -36,16 +41,8 @@ public class MatchApplication {
   @Column(name = "STATUS", nullable = false, columnDefinition = "VARCHAR(20) DEFAULT '대기중'")
   private MatchApplicationStatus status = MatchApplicationStatus.PENDING;
 
-  @CreatedDate
-  @Column(name = "APPLIED_AT", nullable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-  private LocalDateTime appliedAt;
-
   @Column(name = "RESPONDED_AT")
   private LocalDateTime respondedAt;
-
-  @LastModifiedDate
-  @Column(name = "UPDATED_AT", nullable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP")
-  private LocalDateTime updatedAt;
 
   protected MatchApplication() {
   }
@@ -56,9 +53,7 @@ public class MatchApplication {
       this.targetTeam = targetTeam;
       this.applicationMessage = applicationMessage;
       this.status = MatchApplicationStatus.PENDING;
-      this.appliedAt = LocalDateTime.now();
       this.respondedAt = null;
-      this.updatedAt = LocalDateTime.now();
   }
 
   public Long getApplicationId() {
@@ -81,15 +76,7 @@ public class MatchApplication {
     return status;
   }
 
-  public LocalDateTime getAppliedAt() {
-    return appliedAt;
-  }
-
   public LocalDateTime getRespondedAt() {return respondedAt;}
-
-  public LocalDateTime getUpdatedAt() {
-    return updatedAt;
-  }
 
   public MatchQueue getMatchQueue(){ return matchQueue; }
