+++ src/main/java/com/shootdoori/match/entity/Venue.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "venues")
-public class Venue {
+public class Venue extends DateEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -38,12 +38,6 @@ public class Venue {
 
   @Column(name = "PRICE_PER_HOUR")
   private Long pricePerHour;
-
-  @Column(name = "CREATED_AT", updatable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-  private LocalDateTime createdAt;
-
-  @Column(name = "UPDATED_AT", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP")
-  private LocalDateTime updatedAt;
 
   protected Venue() {}
 
@@ -77,13 +71,5 @@ public class Venue {
 
   public Long getPricePerHour() {
     return pricePerHour;
-  }
-
-  public LocalDateTime getCreatedAt() {
-    return createdAt;
-  }
-
-  public LocalDateTime getUpdatedAt() {
-    return updatedAt;
   }
 }
